Refactor HFE implementation into separate module
- Extract HFE loading and saving routines from disc.js into disc-hfe.js
- Create test-disc-hfe.js with HFE tests extracted from test-disc.js
- Update imports/exports in disc.js to maintain backward compatibility
- All tests passing after refactoring
