+++ tests/unit/test-disc-hfe.js
@@ -1,0 +1,303 @@
+import { describe, it } from "vitest";
+import assert from "assert";
+
+import { Disc, DiscConfig, IbmDiscFormat, loadSsd } from "../../src/disc.js";
+import { loadHfe, toHfe } from "../../src/disc-hfe.js";
+import * as fs from "node:fs";
+
+describe("HFE loader tests", function () {
+    const data = fs.readFileSync("public/discs/elite.hfe");
+    it("should load Elite", () => {
+        const disc = new Disc(true, new DiscConfig(), "test.hfe");
+        loadHfe(disc, data);
+        assert.equal(disc.tracksUsed, 81);
+        const sectors = disc.getTrack(false, 0).findSectors();
+        assert.equal(sectors.length, 10);
+        for (const sector of sectors) {
+            assert(!sector.hasHeaderCrcError);
+            assert(!sector.hasDataCrcError);
+        }
+    });
+
+    it("should reject invalid HFE files", () => {
+        const disc = new Disc(true, new DiscConfig(), "test.hfe");
+
+        // Test missing header
+        assert.throws(() => {
+            loadHfe(disc, new Uint8Array(10));
+        }, /HFE file missing header/);
+
+        // Test invalid header
+        const invalidHeader = new Uint8Array(512);
+        invalidHeader.set(new TextEncoder().encode("INVALID!"), 0);
+        assert.throws(() => {
+            loadHfe(disc, invalidHeader);
+        }, /HFE file bad header/);
+
+        // Test non-zero revision
+        const nonZeroRevision = new Uint8Array(512);
+        nonZeroRevision.set(new TextEncoder().encode("HXCHFEV3"), 0);
+        nonZeroRevision[8] = 1; // Set revision to 1 (should be 0)
+        assert.throws(() => {
+            loadHfe(disc, nonZeroRevision);
+        }, /HFE file revision not 0/);
+
+        // Test unsupported encoding
+        const unsupportedEncoding = new Uint8Array(512);
+        unsupportedEncoding.set(new TextEncoder().encode("HXCHFEV3"), 0);
+        unsupportedEncoding[8] = 0; // Revision 0
+        unsupportedEncoding[11] = 1; // Encoding 1 (not 0 or 2)
+        assert.throws(() => {
+            loadHfe(disc, unsupportedEncoding);
+        }, /HFE encoding not ISOIBM/);
+    });
+});
+
+describe(
+    "HFE round-trip tests",
+    {
+        timeout: 120000, // HFE processing can be slow
+    },
+    function () {
+        const data = fs.readFileSync("public/discs/elite.hfe");
+        it("should round-trip elite.hfe", () => {
+            // Load the original HFE file
+            const disc = new Disc(true, new DiscConfig(), "test.hfe");
+            loadHfe(disc, data);
+
+            // Export it back to HFE
+            const hfeSaved = toHfe(disc);
+
+            // Load the saved HFE into a new disc
+            const disc2 = new Disc(true, new DiscConfig(), "test2.hfe");
+            loadHfe(disc2, hfeSaved);
+
+            // Verify that both discs have the same properties
+            assert.equal(disc.tracksUsed, disc2.tracksUsed);
+            assert.equal(disc.isDoubleSided, disc2.isDoubleSided);
+
+            // Compare sectors in a few sample tracks
+            const trackSamples = [0, 10, 20, 40]; // Sample a few tracks
+            for (const trackNum of trackSamples) {
+                if (trackNum >= disc.tracksUsed) continue;
+
+                const track1 = disc.getTrack(false, trackNum);
+                const track2 = disc2.getTrack(false, trackNum);
+
+                // With our variable track length HFE implementation,
+                // track lengths should be identical after roundtripping
+
+                // Track lengths should be identical when roundtripping with the variable track length HFE implementation
+                assert.equal(
+                    track1.length,
+                    track2.length,
+                    `Track ${trackNum} length mismatch: original ${track1.length}, round-tripped ${track2.length}`,
+                );
+
+                // Compare sectors - this is the most important test
+                // All sectors must be readable and contain the correct data
+                const sectors1 = track1.findSectors();
+                const sectors2 = track2.findSectors();
+
+                // All sectors must be found
+                assert.equal(
+                    sectors1.length,
+                    sectors2.length,
+                    `Track ${trackNum} sector count mismatch: ${sectors1.length} vs ${sectors2.length}`,
+                );
+
+                // Compare sector data for first sector as a sample
+                if (sectors1.length > 0 && sectors2.length > 0) {
+                    assert.equal(sectors1[0].sectorNumber, sectors2[0].sectorNumber);
+                    assert.equal(sectors1[0].trackNumber, sectors2[0].trackNumber);
+
+                    // Compare actual sector data if available
+                    if (sectors1[0].sectorData && sectors2[0].sectorData) {
+                        assert.equal(
+                            sectors1[0].sectorData.length,
+                            sectors2[0].sectorData.length,
+                            `Track ${trackNum} sector data length mismatch`,
+                        );
+
+                        // Sample a few bytes from the sector
+                        if (sectors1[0].sectorData.length > 0) {
+                            assert.equal(
+                                sectors1[0].sectorData[0],
+                                sectors2[0].sectorData[0],
+                                `Track ${trackNum} first byte mismatch`,
+                            );
+
+                            const midPoint = Math.floor(sectors1[0].sectorData.length / 2);
+                            assert.equal(
+                                sectors1[0].sectorData[midPoint],
+                                sectors2[0].sectorData[midPoint],
+                                `Track ${trackNum} middle byte mismatch`,
+                            );
+
+                            assert.equal(
+                                sectors1[0].sectorData[sectors1[0].sectorData.length - 1],
+                                sectors2[0].sectorData[sectors2[0].sectorData.length - 1],
+                                `Track ${trackNum} last byte mismatch`,
+                            );
+                        }
+                    }
+                }
+            }
+        });
+    },
+);
+
+describe("HFE export tests", function () {
+    it("should export a simple single-sided disc", { timeout: 10000 }, () => {
+        const disc = new Disc(true, new DiscConfig(), "test.hfe");
+        const sectorData = new Uint8Array(256);
+        sectorData.fill(0xa5);
+
+        // Create a simple FM track
+        const builder = disc.buildTrack(false, 0);
+        builder
+            .appendRepeatFmByte(0xff, IbmDiscFormat.stdGap1FFs)
+            .appendRepeatFmByte(0x00, IbmDiscFormat.stdSync00s)
+            .resetCrc()
+            .appendFmDataAndClocks(IbmDiscFormat.idMarkDataPattern, IbmDiscFormat.markClockPattern)
+            .appendFmByte(0) // track
+            .appendFmByte(0)
+            .appendFmByte(0) // sector
+            .appendFmByte(1)
+            .appendCrc(false)
+            .appendRepeatFmByte(0xff, IbmDiscFormat.stdGap2FFs)
+            .appendRepeatFmByte(0x00, IbmDiscFormat.stdSync00s)
+            .resetCrc()
+            .appendFmDataAndClocks(IbmDiscFormat.dataMarkDataPattern, IbmDiscFormat.markClockPattern)
+            .appendFmChunk(sectorData)
+            .appendCrc(false)
+            .fillFmByte(0xff);
+
+        const hfeData = toHfe(disc);
+
+        // Verify HFE header
+        const header = new TextDecoder("ascii").decode(hfeData.slice(0, 8));
+        assert.equal(header, "HXCHFEV3");
+        assert.equal(hfeData[8], 0); // Revision
+        assert.equal(hfeData[9], 1); // Number of tracks
+        assert.equal(hfeData[10], 1); // Number of sides
+        assert.equal(hfeData[11], 2); // ISOIBM_FM_MFM_ENCODING
+    });
+
+    it("should export and reload the same disc", { timeout: 30000 }, () => {
+        // Create a disc with FM data
+        const disc = new Disc(true, new DiscConfig(), "test.hfe");
+        const data = new Uint8Array(10240); // 10 sectors worth
+        for (let i = 0; i < data.length; i++) {
+            data[i] = i & 0xff;
+        }
+        loadSsd(disc, data, false);
+
+        // Export to HFE
+        const hfeData = toHfe(disc);
+
+        // Reload from HFE
+        const disc2 = new Disc(true, new DiscConfig(), "test2.hfe");
+        loadHfe(disc2, hfeData);
+
+        // Compare track data
+        assert.equal(disc.tracksUsed, disc2.tracksUsed);
+        for (let trackNum = 0; trackNum < disc.tracksUsed; trackNum++) {
+            const track1 = disc.getTrack(false, trackNum);
+            const track2 = disc2.getTrack(false, trackNum);
+
+            // With our improved implementation, track lengths should be exactly the same
+            // when round-tripping through the HFE format
+            assert.equal(
+                track1.length,
+                track2.length,
+                `Track ${trackNum} length mismatch: original ${track1.length}, round-tripped ${track2.length}`,
+            );
+
+            // Find sectors and compare
+            const sectors1 = track1.findSectors();
+            const sectors2 = track2.findSectors();
+
+            // All sectors must be found - this is the critical test
+            assert.equal(
+                sectors1.length,
+                sectors2.length,
+                `Track ${trackNum} has ${sectors1.length} original sectors but ${sectors2.length} after round-trip`,
+            );
+
+            assert.equal(sectors1.length, sectors2.length);
+
+            for (let i = 0; i < sectors1.length; i++) {
+                assert.equal(sectors1[i].sectorNumber, sectors2[i].sectorNumber);
+                assert.equal(sectors1[i].trackNumber, sectors2[i].trackNumber);
+                // Compare sector data if available
+                if (sectors1[i].sectorData && sectors2[i].sectorData) {
+                    assert.deepEqual(sectors1[i].sectorData, sectors2[i].sectorData);
+                }
+            }
+        }
+    });
+
+    it("should export a double-sided disc", { timeout: 10000 }, () => {
+        const disc = new Disc(true, new DiscConfig(), "test.hfe");
+        const data = new Uint8Array(10240); // 10 sectors worth
+        data.fill(0xbb);
+        loadSsd(disc, data, true); // Load as DSD (double-sided)
+
+        const hfeData = toHfe(disc);
+
+        // Verify HFE header
+        const header = new TextDecoder("ascii").decode(hfeData.slice(0, 8));
+        assert.equal(header, "HXCHFEV3");
+        assert.equal(hfeData[10], 2); // Number of sides
+
+        // Reload and verify
+        const disc2 = new Disc(true, new DiscConfig(), "test2.hfe");
+        loadHfe(disc2, hfeData);
+
+        assert(disc2.isDoubleSided);
+        assert.equal(disc.tracksUsed, disc2.tracksUsed);
+    });
+
+    it("should properly handle MFM data", { timeout: 10000 }, () => {
+        const disc = new Disc(true, new DiscConfig(), "test.hfe");
+        const sectorData = new Uint8Array(256);
+        for (let i = 0; i < sectorData.length; i++) {
+            sectorData[i] = (i * 0x11) & 0xff;
+        }
+
+        // Create an MFM track
+        const builder = disc.buildTrack(false, 0);
+        builder
+            .appendRepeatMfmByte(0x4e, 60)
+            .appendRepeatMfmByte(0x00, 12)
+            .resetCrc()
+            .appendMfm3xA1Sync()
+            .appendMfmByte(IbmDiscFormat.idMarkDataPattern)
+            .appendMfmByte(0) // track
+            .appendMfmByte(0)
+            .appendMfmByte(0) // sector
+            .appendMfmByte(1)
+            .appendCrc(true)
+            .appendRepeatMfmByte(0x4e, 22)
+            .appendRepeatMfmByte(0x00, 12)
+            .resetCrc()
+            .appendMfm3xA1Sync()
+            .appendMfmByte(IbmDiscFormat.dataMarkDataPattern)
+            .appendMfmChunk(sectorData)
+            .appendCrc(true)
+            .appendRepeatMfmByte(0x4e, 24)
+            .fillMfmByte(0x4e);
+
+        const hfeData = toHfe(disc);
+
+        // Reload and verify
+        const disc2 = new Disc(true, new DiscConfig(), "test2.hfe");
+        loadHfe(disc2, hfeData);
+
+        const sectors = disc2.getTrack(false, 0).findSectors();
+        assert.equal(sectors.length, 1);
+        assert(sectors[0].isMfm);
+        assert.deepEqual(sectors[0].sectorData, sectorData);
+    });
+});
